Migrating Results Table into its own file
