+++ src/components/ResultsView.tsx
@@ -1,12 +1,8 @@
 import React, { useEffect, useMemo, useRef, useState } from "react";
-import getRoamUrl from "roamjs-components/dom/getRoamUrl";
-import openBlockInSidebar from "roamjs-components/writes/openBlockInSidebar";
 import {
   Button,
   Icon,
-  IconName,
   Tooltip,
-  HTMLTable,
   InputGroup,
   Popover,
   Menu,
@@ -14,20 +10,20 @@ import {
   Switch,
   Intent,
 } from "@blueprintjs/core";
-import { IconNames } from "@blueprintjs/icons";
-import getPageTitleByPageUid from "roamjs-components/queries/getPageTitleByPageUid";
 import getBasicTreeByParentUid from "roamjs-components/queries/getBasicTreeByParentUid";
-import Filter, { Filters } from "roamjs-components/components/Filter";
+import { Filters } from "roamjs-components/components/Filter";
 import getSubTree from "roamjs-components/util/getSubTree";
 import deleteBlock from "roamjs-components/writes/deleteBlock";
 import createBlock from "roamjs-components/writes/createBlock";
-import extractTag from "roamjs-components/util/extractTag";
 import Export from "./Export";
 import parseQuery from "../utils/parseQuery";
 import { getDatalogQuery } from "../utils/fireQuery";
-import getCurrentUserUid from "roamjs-components/queries/getCurrentUserUid";
 import type { Result } from "roamjs-components/types/query-builder";
-import parseResultSettings from "../utils/parseResultSettings";
+import parseResultSettings, {
+  FilterData,
+  Sorts,
+  Views,
+} from "../utils/parseResultSettings";
 import { useExtensionAPI } from "roamjs-components/components/ExtensionApiContext";
 import postProcessResults from "../utils/postProcessResults";
 import setInputSetting from "roamjs-components/util/setInputSetting";
@@ -41,11 +37,7 @@ import { ExportTypes } from "../utils/types";
 import updateBlock from "roamjs-components/writes/updateBlock";
 import getFirstChildUidByBlockUid from "roamjs-components/queries/getFirstChildUidByBlockUid";
 import { Condition } from "../utils/types";
-import { BLOCK_REF_REGEX } from "roamjs-components/dom/constants";
-import getTextByBlockUid from "roamjs-components/queries/getTextByBlockUid";
-
-type Sorts = { key: string; descending: boolean }[];
-type FilterData = Record<string, Filters>;
+import ResultsTable from "./ResultsTable";
 
 const VIEWS: Record<string, { value: boolean }> = {
   link: { value: false },
@@ -54,273 +46,7 @@ const VIEWS: Record<string, { value: boolean }> = {
   alias: { value: true },
 };
 
-const ResultHeader = ({
-  c,
-  results,
-  activeSort,
-  setActiveSort,
-  filters,
-  setFilters,
-  initialFilter,
-}: {
-  c: string;
-  results: Result[];
-  activeSort: Sorts;
-  setActiveSort: (s: Sorts) => void;
-  filters: FilterData;
-  setFilters: (f: FilterData) => void;
-  initialFilter: Filters;
-}) => {
-  const filterData = useMemo(
-    () => ({
-      values: Array.from(new Set(results.map((r) => toCellValue(r[c])))),
-    }),
-    [results, c]
-  );
-  const sortIndex = activeSort.findIndex((s) => s.key === c);
-  return (
-    <td
-      style={{
-        cursor: "pointer",
-        textTransform: "capitalize",
-      }}
-      key={c}
-      onClick={() => {
-        if (sortIndex >= 0) {
-          if (activeSort[sortIndex].descending) {
-            setActiveSort(activeSort.filter((s) => s.key !== c));
-          } else {
-            setActiveSort(
-              activeSort.map((s) =>
-                s.key === c ? { key: c, descending: true } : s
-              )
-            );
-          }
-        } else {
-          setActiveSort([...activeSort, { key: c, descending: false }]);
-        }
-      }}
-    >
-      <div className="flex items-center">
-        <span className="inline-block mr-4">{c}</span>
-        <span>
-          <Filter
-            data={filterData}
-            initialValue={initialFilter}
-            onChange={(newFilters) =>
-              setFilters({ ...filters, [c]: newFilters })
-            }
-            renderButtonText={(s) =>
-              s ? s.toString() : <i style={{ opacity: 0.5 }}>(Empty)</i>
-            }
-            small
-          />
-          {sortIndex >= 0 && (
-            <span>
-              <Icon
-                icon={
-                  activeSort[sortIndex].descending ? "sort-desc" : "sort-asc"
-                }
-              />
-              <span style={{ fontSize: 8 }}>({sortIndex + 1})</span>
-            </span>
-          )}
-        </span>
-      </div>
-    </td>
-  );
-};
-
-const CellEmbed = ({ uid }: { uid: string }) => {
-  const title = getPageTitleByPageUid(uid);
-  const isPage = !!title ? "page-embed" : "block-embed";
-  const contentRef = useRef(null);
-  useEffect(() => {
-    const el = contentRef.current;
-    if (el) {
-      window.roamAlphaAPI.ui.components.renderBlock({
-        uid,
-        el,
-      });
-    }
-  }, [contentRef]);
-  return (
-    <div className="roamjs-query-embed">
-      {isPage === "page-embed" ? (
-        <h1 className="rm-page-title">{title}</h1>
-      ) : (
-        ""
-      )}
-      <div ref={contentRef} className={isPage} />
-    </div>
-  );
-};
-
-const ResultView = ({
-  r,
-  ctrlClick,
-  views,
-  extraColumn,
-  onRefresh,
-}: {
-  r: Result;
-  ctrlClick?: (e: Result) => void;
-  views: { column: string; mode: string; value: string }[];
-  extraColumn?: { row: (e: Result) => React.ReactNode; reserved: RegExp[] };
-  onRefresh: () => void;
-}) => {
-  const rowCells = Object.keys(r).filter(
-    (k) =>
-      !META_REGEX.test(k) &&
-      !(extraColumn && extraColumn.reserved.some((t) => t.test(k)))
-  );
-  const namespaceSetting = useMemo(
-    () =>
-      (
-        window.roamAlphaAPI.data.fast.q(
-          `[:find [pull ?u [:user/settings]] :where [?u :user/uid "${getCurrentUserUid()}"]]`
-        )?.[0]?.[0] as {
-          ":user/settings": {
-            ":namespace-options": { name: "partial" | "none" | "full" }[];
-          };
-        }
-      )?.[":user/settings"]?.[":namespace-options"]?.[0]?.name,
-    []
-  );
-  const cell = (key: string) => {
-    const value = toCellValue(r[`${key}-display`] || r[key] || "");
-    const action = r[`${key}-action`];
-    if (typeof action === "string") {
-      const buttonProps =
-        value.toUpperCase().replace(/\s/g, "_") in IconNames
-          ? { icon: value as IconName }
-          : { text: value };
-      return (
-        <Button
-          {...buttonProps}
-          onClick={() => {
-            document.dispatchEvent(
-              new CustomEvent("roamjs:query-builder:action", {
-                detail: {
-                  action,
-                  uid: r[`${key}-uid`],
-                  val: r["text"],
-                  onRefresh,
-                },
-              })
-            );
-          }}
-        />
-      );
-    }
-    const formattedValue =
-      typeof value === "string" &&
-      r[`${key}-uid`] &&
-      !!getPageTitleByPageUid((r[`${key}-uid`] || "").toString())
-        ? namespaceSetting === "full"
-          ? value.split("/").slice(-1)[0]
-          : namespaceSetting === "partial"
-          ? value
-              .split("/")
-              .map((v, i, a) => (i === a.length - 1 ? v : v.slice(0, 1)))
-              .join("/")
-          : value
-        : value;
-    return formattedValue
-      .toString()
-      .split("<span>")
-      .map((s, i) => (
-        <span
-          key={i}
-          className={i % 2 === 0 ? "" : "roamjs-query-hightlighted-result"}
-        >
-          {s}
-        </span>
-      ));
-  };
-  const viewsByColumn = useMemo(
-    () => Object.fromEntries(views.map((v) => [v.column, v])),
-    [views]
-  );
-  return (
-    <>
-      <tr>
-        {rowCells.map((k) => {
-          const uid = (r[`${k}-uid`] || "").toString();
-          const val = r[k] || "";
-          const { mode: view, value: viewValue } = viewsByColumn[k] || {};
-          return (
-            <td
-              style={{
-                textOverflow: "ellipsis",
-                overflow: "hidden",
-              }}
-              key={k}
-            >
-              {val === "" ? (
-                <i>[block is blank]</i>
-              ) : view === "link" || view === "alias" ? (
-                <a
-                  className={"rm-page-ref"}
-                  data-link-title={getPageTitleByPageUid(uid) || ""}
-                  href={getRoamUrl(uid)}
-                  onMouseDown={(e) => {
-                    if (e.shiftKey) {
-                      openBlockInSidebar(uid);
-                      e.preventDefault();
-                      e.stopPropagation();
-                    } else if (e.ctrlKey) {
-                      ctrlClick?.({
-                        text: toCellValue(val),
-                        uid,
-                      });
-                      e.preventDefault();
-                      e.stopPropagation();
-                    }
-                  }}
-                  onClick={(e) => {
-                    if (e.shiftKey || e.ctrlKey) {
-                      e.preventDefault();
-                      e.stopPropagation();
-                    }
-                  }}
-                  onContextMenu={(e) => {
-                    if (e.ctrlKey) {
-                      e.preventDefault();
-                      e.stopPropagation();
-                    }
-                  }}
-                >
-                  {view === "alias" ? viewValue : cell(k)}
-                </a>
-              ) : view === "embed" ? (
-                <CellEmbed uid={uid} />
-              ) : (
-                cell(k)
-              )}
-            </td>
-          );
-        })}
-        {extraColumn && <td>{extraColumn.row(r)}</td>}
-      </tr>
-    </>
-  );
-};
-
-const resolveRefs = (text: string, refs = new Set<string>()): string => {
-  return text.replace(new RegExp(BLOCK_REF_REGEX, "g"), (_, blockUid) => {
-    if (refs.has(blockUid)) return "";
-    const reference = getTextByBlockUid(blockUid);
-    return resolveRefs(reference, new Set(refs));
-  });
-};
 const META_REGEX = /(-(uid|action|display)$|^uid$)/;
-const toCellValue = (v: number | Date | string) =>
-  v instanceof Date
-    ? window.roamAlphaAPI.util.dateToPageTitle(v)
-    : typeof v === "undefined" || v === null
-    ? ""
-    : extractTag(resolveRefs(v.toString()));
 type EnglishQueryPart = { text: string; clickId?: string };
 
 const QueryUsed = ({ parentUid }: { parentUid: string }) => {
@@ -459,22 +185,23 @@ type ResultsViewComponent = (props: {
   header?: React.ReactNode;
   results: Result[];
   hideResults?: boolean;
-  ctrlClick?: (e: Result) => void;
   preventSavingSettings?: boolean;
   preventExport?: boolean;
   onEdit?: () => void;
   onRefresh: () => void;
   getExportTypes?: (r: Result[]) => ExportTypes;
-  onResultsInViewChange?: (r: Result[]) => void;
   globalFiltersData?: Record<string, Filters>;
   globalPageSize?: number;
+  isEditBlock?: boolean;
+  // @deprecated - should be inferred from the query or layout
+  onResultsInViewChange?: (r: Result[]) => void;
+  // @deprecated - move to `-action` columns through selections
   extraColumn?: {
     reserved: RegExp[];
     width: number;
     row: (e: Result) => React.ReactNode;
     header: React.ReactNode;
   };
-  isEditBlock?: boolean;
 }) => JSX.Element;
 
 const ResultsView: ResultsViewComponent = ({
@@ -482,15 +209,14 @@ const ResultsView: ResultsViewComponent = ({
   header,
   results,
   hideResults = false,
-  ctrlClick,
   preventSavingSettings = false,
   preventExport,
   onEdit,
   onRefresh,
   getExportTypes,
   onResultsInViewChange,
-  extraColumn,
   isEditBlock,
+  extraColumn,
 }) => {
   const extensionAPI = useExtensionAPI();
   const columns = useMemo(
@@ -518,7 +244,7 @@ const ResultsView: ResultsViewComponent = ({
   const [page, setPage] = useState(settings.page);
   const [pageSize, setPageSize] = useState(settings.pageSize);
   const pageSizeTimeoutRef = useRef(0);
-  const [views, setViews] = useState(settings.views);
+  const [views, setViews] = useState<Views>(settings.views);
   const [searchFilter, setSearchFilter] = useState(() => settings.searchFilter);
 
   const { allResults, paginatedResults } = useMemo(() => {
@@ -944,224 +670,23 @@ const ResultsView: ResultsViewComponent = ({
         {!hideResults &&
           (results.length !== 0 ? (
             layout === "table" ? (
-              <HTMLTable
-                style={{
-                  maxHeight: "400px",
-                  overflowY: "scroll",
-                  width: "100%",
-                  tableLayout: "fixed",
-                  borderRadius: 3,
-                }}
-                striped
-                interactive
-                bordered
-              >
-                <thead
-                  style={{
-                    background: "#eeeeee80",
-                  }}
-                >
-                  <tr>
-                    {columns.map((c) => (
-                      <ResultHeader
-                        key={c}
-                        c={c}
-                        results={results}
-                        activeSort={activeSort}
-                        setActiveSort={(as) => {
-                          setActiveSort(as);
-                          if (preventSavingSettings) return;
-                          const resultNode = getSubTree({
-                            key: "results",
-                            parentUid,
-                          });
-                          const sortsNode = getSubTree({
-                            key: "sorts",
-                            parentUid: resultNode.uid,
-                          });
-                          sortsNode.children.forEach((c) => deleteBlock(c.uid));
-                          as.map((a) => ({
-                            text: a.key,
-                            children: [{ text: `${a.descending}` }],
-                          })).forEach((node, order) =>
-                            createBlock({
-                              parentUid: sortsNode.uid,
-                              node,
-                              order,
-                            })
-                          );
-                        }}
-                        filters={filters}
-                        setFilters={(fs) => {
-                          setFilters(fs);
-
-                          if (preventSavingSettings) return;
-                          const resultNode = getSubTree({
-                            key: "results",
-                            parentUid,
-                          });
-                          const filtersNode = getSubTree({
-                            key: "filters",
-                            parentUid: resultNode.uid,
-                          });
-                          filtersNode.children.forEach((c) =>
-                            deleteBlock(c.uid)
-                          );
-                          Object.entries(fs)
-                            .filter(
-                              ([, data]) =>
-                                data.includes.values.size ||
-                                data.excludes.values.size
-                            )
-                            .map(([column, data]) => ({
-                              text: column,
-                              children: [
-                                {
-                                  text: "includes",
-                                  children: Array.from(
-                                    data.includes.values
-                                  ).map((text) => ({ text })),
-                                },
-                                {
-                                  text: "excludes",
-                                  children: Array.from(
-                                    data.excludes.values
-                                  ).map((text) => ({ text })),
-                                },
-                              ],
-                            }))
-                            .forEach((node, order) =>
-                              createBlock({
-                                parentUid: filtersNode.uid,
-                                node,
-                                order,
-                              })
-                            );
-                        }}
-                        initialFilter={settings.filters[c]}
-                      />
-                    ))}
-                    {extraColumn && (
-                      <th style={{ width: extraColumn.width }}>
-                        {extraColumn.header}
-                      </th>
-                    )}
-                  </tr>
-                </thead>
-                <tbody>
-                  {paginatedResults.map((r) => (
-                    <ResultView
-                      key={Object.values(r).join("-")}
-                      r={r}
-                      ctrlClick={ctrlClick}
-                      views={views}
-                      extraColumn={extraColumn}
-                      onRefresh={onRefresh}
-                    />
-                  ))}
-                </tbody>
-                <tfoot>
-                  <tr>
-                    <td
-                      colSpan={columns.length + (extraColumn ? 1 : 0)}
-                      style={{ padding: 0, background: "#eeeeee80" }}
-                    >
-                      <div
-                        className="flex justify-between items-center"
-                        style={{
-                          opacity: 0.8,
-                          padding: "8px 4px",
-                        }}
-                      >
-                        <span>
-                          <i style={{ opacity: 0.8 }}>
-                            <Button
-                              icon={showContent ? "caret-down" : "caret-right"}
-                              minimal
-                              small
-                              onClick={() => setShowContent(!showContent)}
-                              style={{
-                                marginRight: 4,
-                              }}
-                            />
-                            Showing {paginatedResults.length} of{" "}
-                            {results.length} results
-                          </i>
-                        </span>
-                        <span style={{ display: "flex", alignItems: "center" }}>
-                          <span>Rows per page:</span>
-                          <InputGroup
-                            defaultValue={pageSize.toString()}
-                            onChange={(e) => {
-                              clearTimeout(pageSizeTimeoutRef.current);
-                              pageSizeTimeoutRef.current = window.setTimeout(
-                                () => {
-                                  setPageSize(Number(e.target.value));
-
-                                  if (preventSavingSettings) return;
-                                  const resultNode = getSubTree({
-                                    key: "results",
-                                    parentUid,
-                                  });
-                                  setInputSetting({
-                                    key: "size",
-                                    value: e.target.value,
-                                    blockUid: resultNode.uid,
-                                  });
-                                },
-                                1000
-                              );
-                            }}
-                            type="number"
-                            style={{
-                              width: 60,
-                              maxWidth: 60,
-                              marginRight: 32,
-                              marginLeft: 16,
-                            }}
-                          />
-                          <Button
-                            minimal
-                            icon={"double-chevron-left"}
-                            onClick={() => setPage(1)}
-                            disabled={page === 1}
-                          />
-                          <Button
-                            minimal
-                            icon={"chevron-left"}
-                            onClick={() => setPage(page - 1)}
-                            disabled={page === 1}
-                          />
-                          <span style={{ margin: "4px 0" }}>{page}</span>
-                          <Button
-                            minimal
-                            icon={"chevron-right"}
-                            onClick={() => setPage(page + 1)}
-                            disabled={
-                              page ===
-                                Math.ceil(allResults.length / pageSize) ||
-                              allResults.length === 0
-                            }
-                          />
-                          <Button
-                            minimal
-                            icon={"double-chevron-right"}
-                            disabled={
-                              page ===
-                                Math.ceil(allResults.length / pageSize) ||
-                              allResults.length === 0
-                            }
-                            onClick={() =>
-                              setPage(Math.ceil(allResults.length / pageSize))
-                            }
-                          />
-                        </span>
-                      </div>
-                      {showContent && <QueryUsed parentUid={parentUid} />}
-                    </td>
-                  </tr>
-                </tfoot>
-              </HTMLTable>
+              <ResultsTable
+                columns={columns}
+                results={paginatedResults}
+                parentUid={parentUid}
+                activeSort={activeSort}
+                setActiveSort={setActiveSort}
+                filters={filters}
+                setFilters={setFilters}
+                views={views}
+                page={page}
+                setPage={setPage}
+                pageSize={pageSize}
+                setPageSize={setPageSize}
+                pageSizeTimeoutRef={pageSizeTimeoutRef}
+                onRefresh={onRefresh}
+                allResultsLength={allResults.length}
+              />
             ) : layout === "line" ? (
               <Charts
                 type="line"
@@ -1180,34 +705,33 @@ const ResultsView: ResultsViewComponent = ({
               <p className="px-2 py-3 flex justify-between items-center mb-0">
                 <i>No Results Found</i>
               </p>
-              <div style={{ background: "#eeeeee80" }}>
-                <div
-                  className="flex justify-between items-center"
-                  style={{
-                    opacity: 0.8,
-                    padding: "8px 4px",
-                  }}
-                >
-                  <span>
-                    <i style={{ opacity: 0.8 }}>
-                      <Button
-                        icon={showContent ? "caret-down" : "caret-right"}
-                        minimal
-                        small
-                        onClick={() => setShowContent(!showContent)}
-                        style={{
-                          marginRight: 4,
-                        }}
-                      />
-                      Showing {paginatedResults.length} of {results.length}{" "}
-                      results
-                    </i>
-                  </span>
-                </div>
-                {showContent && <QueryUsed parentUid={parentUid} />}
-              </div>
             </>
           ))}
+        <div style={{ background: "#eeeeee80" }}>
+          <div
+            className="flex justify-between items-center"
+            style={{
+              opacity: 0.8,
+              padding: "8px 4px",
+            }}
+          >
+            <span>
+              <i style={{ opacity: 0.8 }}>
+                <Button
+                  icon={showContent ? "caret-down" : "caret-right"}
+                  minimal
+                  small
+                  onClick={() => setShowContent(!showContent)}
+                  style={{
+                    marginRight: 4,
+                  }}
+                />
+                Showing {paginatedResults.length} of {results.length} results
+              </i>
+            </span>
+          </div>
+          {showContent && <QueryUsed parentUid={parentUid} />}
+        </div>
       </div>
     </div>
   );
